requirements.txt
Flask>=3,<4
requests>=2.31,<3
python-dotenv>=1,<2
gunicorn>=21,<22